Add user account creation and saving to database
Introducing two new functions, [createUserAccount] and [saveUserToDB], to handle the creation of a new user account and saving the user data to the database.

[createUserAccount] takes in a `user` object with properties like `name`, `email`, `password`, and `username`, and creates a new account using the `account.create` method from the `appwrite` library. It also generates an avatar URL and saves the user data to the database using the [saveUserToDB] function.

[saveUserToDB] takes in a `user` object with properties like `accountId`, `email`, `name`, `imageUrl`, and `username`, and saves the user data to the database using the `databases.createDocument` method from the `appwrite` library.

These functions are designed to be used together to handle the complete process of creating a new user account and saving the user data to the database.